Use a relative path to the BLT submodule
this is a special Git trick: if two Git
repositories exist on the same server,
than the submodule can be specified
using a relative path.

This is useful because then the Git
protocol being used to clone the parent
repository will also be used to clone
the submodule.

In my case, I have an environment where
the HTTP protocol has proxy issues but
the SSH protocol has no issues and is
better for scripting.
With this change, recursively cloning
git@github.com:LLNL/Adiak.git
will also use the SSH protocol when
trying to clone BLT, which prevents
it from hitting HTTP issues